requirements_test.txt
#django>=1.8.0
django-model-utils>=2.4
django_markupfield>=1.3.5
markdown>=2.6.5
django-taggit>=0.17.6
django-reversion>=1.10.0
pytz==2015.7
django-haystack>=2.4.1
whoosh>=2.7.0
#
coverage
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
model-mommy>=1.2.6
